requirements.txt
appdirs==1.4.4
arrow==1.3.0
asgiref==3.8.1
beautifulsoup4==4.12.3
binaryornot==0.4.4
bs4==0.0.2
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
cssselect==1.2.0
Django==5.1.1
django-tailwind==3.8.0
djangorestframework==3.15.2
fake-useragent==1.5.1
feedparser==6.0.11
frozendict==2.4.4
html5lib==1.1
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
lxml==5.3.0
lxml_html_clean==0.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multitasking==0.0.11
numpy==2.1.1
pandas==2.2.3
parse==1.20.2
peewee==3.17.6
platformdirs==4.3.6
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyee==11.1.1
Pygments==2.18.0
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-html==0.10.0
rich==13.8.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
text-unidecode==1.3
tqdm==4.66.5
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
w3lib==2.2.1
webencodings==0.5.1
websockets==10.4
yahoo-fin==0.8.9.1
yfinance==0.2.43
zipp==3.20.2
